+++ config/settings.py
@@ -283,9 +283,11 @@ DEFAULT_REQUEST_TIMEOUT = get_env(
     "OSIDB_DEFAULT_REQUEST_TIMEOUT", default="30", is_int=True
 )
 
-# sets the Access-Control-Allow-Origin response header
-# example value: ["https://osidb.example.com", "https://workflows.example.com"]
-CORS_ALLOWED_ORIGINS = get_env("OSIDB_CORS_ALLOWED_ORIGINS", default="[]", is_json=True)
+# sets the Access-Control-Allow-Origin response header - accepts regex
+# example value: [ r"^https://([^.]*\.)?\.example\.com$" ]
+CORS_ALLOWED_ORIGIN_REGEXES = get_env(
+    "OSIDB_CORS_ALLOWED_ORIGINS", default="[]", is_json=True
+)
 # sets the Access-Control-Allow-Headers response header; lowercase
 CORS_ALLOW_HEADERS = (
     *default_headers,